+++ smart-contracts/wasm-transform/CHANGELOG.md
@@ -8,6 +8,8 @@
     initialization sections.
   - allow instructions defined in the [sign extension operators](https://github.com/WebAssembly/sign-extension-ops/blob/master/proposals/sign-extension-ops/Overview.md)
     Wasm proposal.
+- `instantiate` and `instantiate_with_metering`'s return type is changed from
+  `Artifact` to `InstantiatedModule` which adds metadata on top of the artifact.
 
 ## concordium-wasm 2.0.0 (2023-06-16)